+++ db/migrate/20220426114431_remove_links_from_models.rb
@@ -1,0 +1,9 @@
+# frozen_string_literal: true
+
+class RemoveLinksFromModels < ActiveRecord::Migration[6.1]
+  def change
+    remove_column :services, :use_cases_url
+    remove_column :services, :multimedia
+    remove_column :providers, :multimedia
+  end
+end