+++ app/services/importers/service.rb
@@ -96,6 +96,7 @@ class Importers::Service < ApplicationService
       pid: @data["id"],
       # Basic
       name: @data["name"],
+      abbreviation: @data["abbreviation"],
       resource_organisation: map_provider(@data["resourceOrganisation"], @eosc_registry_base_url, token: @token),
       providers: providers.uniq.map { |p| map_provider(p, @eosc_registry_base_url, token: @token) },
       webpage_url: @data["webpage"] || "",