+++ db/migrate/20220524154008_rename_hosting_legal_entity.rb
@@ -1,0 +1,7 @@
+# frozen_string_literal: true
+
+class RenameHostingLegalEntity < ActiveRecord::Migration[6.1]
+  def change
+    rename_column :providers, :hosting_legal_entity, :hosting_legal_entity_string
+  end
+end