+++ app/jobs/vocabulary/pc_create_or_update_job.rb
@@ -1,0 +1,11 @@
+# frozen_string_literal: true
+
+class Vocabulary::PcCreateOrUpdateJob < ApplicationJob
+  queue_as :pc_subscriber
+
+  rescue_from(Errno::ECONNREFUSED) { |exception| raise exception }
+
+  def perform(vocabulary)
+    Vocabulary::PcCreateOrUpdate.new(vocabulary).call
+  end
+end