+++ CHANGELOG.md
@@ -9,7 +9,8 @@ Please view this file on the master branch, on stable branches it's out of date.
 ## [3.38.0-milestone]
 
 ### Added
-- Dependencies for Provider Provile v4.0 (@kmarszalek)
+- Dependencies for Provider Profile v4.0 (@kmarszalek)
+- Catalogue for Resource Profile 4.0 (@goreck888)
 
 ### Changed
 - Move Structure Types from `Other` to `Classification` section in the Resource Profile 4.0 (@goreck888)