+++ CHANGELOG.md
@@ -11,6 +11,7 @@ Please view this file on the master branch, on stable branches it's out of date.
 ### Added
 - Dependencies for Provider Profile v4.0 (@kmarszalek)
 - Catalogue for Resource Profile 4.0 (@goreck888)
+- Abbreviation for the Resource Profile 4.0 (@goreck888)
 
 ### Changed
 - Move Structure Types from `Other` to `Classification` section in the Resource Profile 4.0 (@goreck888)