+++ CHANGELOG.md
@@ -16,8 +16,10 @@ Please view this file on the master branch, on stable branches it's out of date.
 ### Changed
 - Move Structure Types from `Other` to `Classification` section in the Resource Profile 4.0 (@goreck888)
 - `Multimedia` and `UseCases` urls specified in the Resource Profile 4.0 with backward compatibility (@goreck888)
+- Replace `OpenAIRE Explore` banner with the `EOSC Explore` one for `EOSC` tagged resources (@goreck888)
 
 ### Deprecated
+- `OPENAIRE_NOTEBOOKS__HARDCODED_LINK` environmental variable (@goreck888) 
 
 ### Removed