+++ CHANGELOG.md
@@ -26,6 +26,7 @@ Please view this file on the master branch, on stable branches it's out of date.
 ### Removed
 
 ### Fixed
+- JMS Resource update on missing providers (@goreck888)
 
 ### Security